+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.3 - 2018-08-28
+### Added
+- Update actions to work with craft behavior (keep the entry selected after call)
+- Now support radio button
+
 ## 1.0.2 - 2018-08-24
 ### Added
 - Dynamic selection of sections and fields to bulk edit in the elements listing